requirements.txt
fastapi
uvicorn
torch
torchvision
librosa
numpy
pillow
python-multipart
soundfile
